% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default}
\alias{default}
\title{Non-blocking match.arg with better error reporting.}
\usage{
default(arg, choices, message = NULL)
}
\description{
Performs similar traceback as match.arg for default argument values. Unlike
match.arg, this function does not stop execution. It simply emits a detailed
warning including the invalid argument value, the calling funcation, and
the expected argument values.
}
\keyword{internal}
