% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{Shiny Searchbar Demos}
\usage{
demo(name = "gallery", ...)
}
\arguments{
\item{name}{Name of the demo, either "gallery" or "overflow".}

\item{...}{Remaining arguments are passed to \code{shiny::runApp}, e.g.
\code{'port'}, \code{'display.mode'}, etc.}
}
\description{
Run one of the the shinySearchbar demos: "gallery" and "overflow".
}
\examples{
if (interactive()) {
  shinySearchbar:::demo("gallery")
}

}
