% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%_\%}
\alias{\%_\%}
\title{Underscore Join Operator}
\usage{
x \%_\% y
}
\description{
Inspired by Shiny's \code{\%.\%} operator which joins two arguments with a
period (.), instead this uses an underscore (_) which is more suitable
for HTML \code{id} attributes.
}
\keyword{internal}
