% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbar.R
\name{updateMarkOptions}
\alias{updateMarkOptions}
\title{Update mark.js Options for an Initialized Searchbar Widget}
\usage{
updateMarkOptions(
  inputId,
  markOpts,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The \code{input} slot of the initialized searchbar widget.}

\item{markOpts}{Options to update for mark.js API.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
Only the options explicitly passed with this function are updated, the
previous options are left unchanged.
}
\seealso{
\link{configurator}
}
